shrink ch^2 diff with direct calc to 4426 - 4426 == -0.000145

+++ src/GapsParameters.h
@@ -45,10 +45,10 @@ public:
     unsigned outputFrequency;
     unsigned checkpointInterval;
     unsigned snapshotFrequency;
-    float alphaA;
-    float alphaP;
-    float maxGibbsMassA;
-    float maxGibbsMassP;
+    double alphaA;
+    double alphaP;
+    double maxGibbsMassA;
+    double maxGibbsMassP;
     PumpThreshold pumpThreshold;
     GapsAlgorithmPhase snapshotPhase;
     bool useFixedPatterns;
